producer/requirements.txt
requests
avro
confluent_kafka
load-dotenv
datetime
python-dateutil
redis